[refactor] Move FP16LayerPlanner and Q8_0LayerPlanner to quantized-specific subpackages
